+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [2.1.0] - 2026-08-37
 
+- Improvement: Add function to read pyproject file
 - Improvement: Add function to check if module is one file
 
 ## [2.0.1] - 2026-01-31